+++ src/schemas/product.schema.ts
@@ -1,6 +1,7 @@
 import { Schema } from 'mongoose';
+import { Product } from '../models/product.model';
 
-export const productSchema = new Schema(
+export const productSchema: Schema<Product> = new Schema(
     {
         name: String,
         description: String,
